Create SlaveDevice1 (mbed)
This code is to be loaded on an mbed for the purposes of controlling a slot car within a higher system. This program simply outputs a constant PWM signal to an attached L298N board to provide steady speed to a slotcar (powered by the L298N output).
The mbed is also programmed to receive and Interrupt from a Master Device, which triggers a PWM change. Currently the PWM modification only causes a complete stop of the signal, for a specified time. There is scope (with further testing to dynamically reduce speed for a custom amount of time - through i2c).